hey Trait! You Are'nt Gonna Need It :D
